refactor: di and bump deps
