+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/di/CoroutinesDispatchersModule.kt
@@ -1,4 +1,4 @@
-package net.nymtech.nymvpn.module
+package net.nymtech.nymvpn.di
 
 import dagger.Module
 import dagger.Provides
@@ -6,10 +6,10 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
-import net.nymtech.nymvpn.module.qualifiers.DefaultDispatcher
-import net.nymtech.nymvpn.module.qualifiers.IoDispatcher
-import net.nymtech.nymvpn.module.qualifiers.MainDispatcher
-import net.nymtech.nymvpn.module.qualifiers.MainImmediateDispatcher
+import net.nymtech.nymvpn.di.qualifiers.DefaultDispatcher
+import net.nymtech.nymvpn.di.qualifiers.IoDispatcher
+import net.nymtech.nymvpn.di.qualifiers.MainDispatcher
+import net.nymtech.nymvpn.di.qualifiers.MainImmediateDispatcher
 
 @InstallIn(SingletonComponent::class)
 @Module
