+++ nym-vpn-android/gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-accompanist = "0.37.0"
+accompanist = "0.37.2"
 agp = "8.8.0-alpha05"
 coreSplashscreen = "1.2.0-alpha02"
 detektRulesCompose = "1.4.0"
@@ -32,7 +32,7 @@ desugar = "2.1.4"
 zxingAndroidEmbedded = "4.3.0"
 
 gradlePlugins-licensee = "1.12.0"
-gradlePlugins-detekt = "1.23.7"
+gradlePlugins-detekt = "1.23.8"
 gradlePlugins-gross = "0.4.4"
 material = "1.12.0"
 gradlePlugins-ktlint="12.1.2"
